requirements.txt
Flask>=2.0.0
pure-python-adb>=0.3.0.dev0
python-dotenv>=1.0.0
